web/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
gunicorn>=21.2,<22.0
python-dotenv>=1.0,<2.0
msal>=1.28,<2.0
requests>=2.32,<3.0
